Fix rendering of stability pool deposits and earn position hooks
