+++ frontend/app/src/comps/Positions/PositionCardEarn.tsx
@@ -101,7 +101,7 @@ export function PositionCardEarn({
                       <Amount
                         fallback='−'
                         percentage
-                        value={earnPool.data.apr}
+                        value={earnPool.data?.apr}
                       />
                     </div>
                   </div>
@@ -126,7 +126,7 @@ export function PositionCardEarn({
                       <Amount
                         fallback='−'
                         percentage
-                        value={earnPool.data.apr7d}
+                        value={earnPool.data?.apr7d}
                       />
                     </div>
                   </div>
