+++ frontend/app/src/screens/EarnPoolScreen/EarnPoolScreen.tsx
@@ -169,7 +169,7 @@ export function EarnPoolScreen() {
                 {tab.action === "deposit" && (
                   <PanelUpdateDeposit
                     collIndex={collIndex}
-                    deposited={earnPool.data.totalDeposited ?? dn.from(0, 18)}
+                    deposited={earnPool.data?.totalDeposited ?? dn.from(0, 18)}
                     position={earnPosition.data ?? undefined}
                   />
                 )}
